# Store raw headers aside to avoid mutation