My previous contribution led to a case where RawHeader() return is affected by normalization. If normalization is turned off, things like Content-Length get broken.

This patch makes a copy of raw headers to guarantee the RawHeader() contract is unaffected by normalization switch.

Amortized, the performance hit is negligible:

```sh
name                       old time/op    new time/op    delta
ServerPost10KReqPerConn-8     320ns ± 2%     338ns ± 7%  +5.49%  (p=0.003 n=5+10)

name                       old alloc/op   new alloc/op   delta
ServerPost10KReqPerConn-8     0.00B          0.00B         ~     (all equal)

name                       old allocs/op  new allocs/op  delta
ServerPost10KReqPerConn-8      0.00           0.00         ~     (all equal)
```