@illotum 2018-06-15
Existing tests were adjusted to always expect unnormalized headers -- previously they were expecting what the "normalize" flag was set to -- so will fail if headers are changed during normalization.

For example, compare dropped `exp` value from the diff to `kvs` content in "normalized" test in headers_test.go L117.