@erikdubbelboer 2018-06-15
Ok I guess this might be a breaking change for people who expect `RawHeaders()` to return normalized headers. But seeing as this function wasn't included in the original fasthttp I guess it isn't used very much yet. So I decided to merge this. Thanks for the fix.